přidání čísla účtu do profilu, úprava konfiguračního formuláře (povinné zadání čísla účtu, ne jen pro tisk)
