+++ app/BackModule/presenters/ConfigurationPresenter.php
@@ -81,6 +81,31 @@ class ConfigurationPresenter extends BasePresenter
         return new \SRS\Form\Configuration\SettingsForm(null, null, $this->dbsettings, $this->context->parameters);
     }
 
+    protected function createComponentSettingsFormCustom()
+    {
+        return new \SRS\Form\Configuration\SettingsFormCustom(null, null, $this->dbsettings, $this->context->parameters);
+    }
+
+    protected function createComponentSettingsFormPayment()
+    {
+        return new \SRS\Form\Configuration\SettingsFormPayment(null, null, $this->dbsettings);
+    }
+
+    protected function createComponentSettingsFormPrint()
+    {
+        return new \SRS\Form\Configuration\SettingsFormPrint(null, null, $this->dbsettings);
+    }
+
+    protected function createComponentSettingsFormProgram()
+    {
+        return new \SRS\Form\Configuration\SettingsFormProgram(null, null, $this->dbsettings);
+    }
+
+    protected function createComponentSettingsFormSeminar()
+    {
+        return new \SRS\Form\Configuration\SettingsFormSeminar(null, null, $this->dbsettings);
+    }
+
     protected function createComponentSkautISEventForm()
     {
         try {
