+++ app/FrontModule/templates/Profile/default.latte
@@ -89,6 +89,11 @@
                     </tr>
 
                     <tr>
+                        <td><strong>Uhraďte na účet</strong></td>
+                        <td>{$accountNumber}</td>
+                    </tr>
+
+                    <tr>
                         <td><strong>Poplatek uhrazen dne </strong></td>
                         <td>
                             {if $dbuser->paymentDate !== null}
